Fixed a lot of signed and unsigned integer comparisons

+++ include/eigen_utils.h
@@ -340,7 +340,7 @@ struct MatrixHash : std::unary_function<T, size_t> {
         // row-major). It will give you the same hash value for two different matrices if they
         // are the transpose of each other in different storage order.
         size_t seed = 0;
-        for (size_t i = 0; i < matrix.size(); ++i) {
+        for (int i = 0; i < matrix.size(); ++i) {
         auto elem = *(matrix.data() + i);
         seed ^= std::hash<typename T::Scalar>()(elem) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
         }
